Standardize AwsLambda The `aws_lambda.py` is not the standard file path. The `aws_` prefix is not needed. For the file name I used the same path as the hook.
